Move Prism::RipperCompat to Prism::Translation::Ripper
